Fix duplicate prebuilt indexes with and without postpended md5 (#2413)

+++ src/main/java/io/anserini/util/PrebuiltIndexHandler.java
@@ -31,7 +31,7 @@ import java.net.HttpURLConnection;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 
 public class PrebuiltIndexHandler {
   private String indexName;
@@ -51,8 +51,8 @@ public class PrebuiltIndexHandler {
      * Get the pyserini cache path firs to avoid double downloads. If the pyserini
      * cache path does not exist, use the anserini cache path.
      */
-    final Path PyseriniPath = Paths.get(System.getProperty("user.home"), ".cache", "pyserini", "indexes");
-    final Path AnseriniPath = Paths.get(System.getProperty("user.home"), ".cache", "anserini", "indexes");
+    final Path PyseriniPath = Path.of(System.getProperty("user.home"), ".cache", "pyserini", "indexes");
+    final Path AnseriniPath = Path.of(System.getProperty("user.home"), ".cache", "anserini", "indexes");
     if (checkFileExist(PyseriniPath)) {
       return PyseriniPath.toString();
     } else {
@@ -69,8 +69,11 @@ public class PrebuiltIndexHandler {
      * Check if the index file exists. If the index file exists, return true.
      * Otherwise, return false.
      */
-    if (checkFileExist(savePath) || checkFileExist(Paths.get(savePath.toString().replace(".gz", "")))
-        || checkFileExist(Paths.get(savePath.toString().replace(".tar.gz", "")))) {
+    if (checkFileExist(savePath) || checkFileExist(Path.of(savePath.toString().replace(".gz", "")))
+        || checkFileExist(Path.of(savePath.toString().replace(".tar.gz", "")))
+        || checkFileExist(Path.of(savePath.toString() + "." + info.md5))
+        || checkFileExist(Path.of(savePath.toString().replace(".gz", "") + "." + info.md5))
+        || checkFileExist(Path.of(savePath.toString().replace(".tar.gz", "") + "." + info.md5))) {
       return true;
     }
     return false;
@@ -102,14 +105,14 @@ public class PrebuiltIndexHandler {
     }
     info = getIndexInfo(indexName);
     // check if saveRootPath exists
-    if (!checkFileExist(Paths.get(saveRootPath))) {
+    if (!checkFileExist(Path.of(saveRootPath))) {
       try {
-        Files.createDirectories(Paths.get(saveRootPath));
+        Files.createDirectories(Path.of(saveRootPath));
       } catch (IOException e) {
         e.printStackTrace();
       }
     }
-    savePath = Paths.get(saveRootPath, info.filename);
+    savePath = Path.of(saveRootPath, info.filename);
     initialized = true;
   }
 
@@ -173,19 +176,19 @@ public class PrebuiltIndexHandler {
     }
 
     String indexFolder = savePath.toString().replace(".tar.gz", "");
-    if (checkFileExist(Paths.get(indexFolder))) {
+    if (checkFileExist(Path.of(indexFolder + "." + info.md5))) {
       System.out.println("Index folder already exists!");
-      return indexFolder;
+      return indexFolder + "." + info.md5;
     }
     System.out.println("Decompressing index...");
 
-    if (checkFileExist(Paths.get(savePath.toString()))) {
+    if (checkFileExist(Path.of(savePath.toString()))) {
       ProcessBuilder pbGZIP = new ProcessBuilder("gzip", "-d", savePath.toString());
       Process pGZIP = pbGZIP.start();
       pGZIP.waitFor();
     }
 
-    if (checkFileExist(Paths.get(savePath.toString().replace(".gz", "")))) {
+    if (checkFileExist(Path.of(savePath.toString().replace(".gz", "")))) {
       ProcessBuilder pbTAR = new ProcessBuilder("tar", "-xvf",
           savePath.toString().substring(0, savePath.toString().length() - 3), "-C", saveRootPath);
       Process pTar = pbTAR.start();
@@ -198,10 +201,12 @@ public class PrebuiltIndexHandler {
     System.out.println("Index decompressed successfully!");
 
     // postpend md5 to decompressed index
-    Path oldIndexPath = Paths.get(indexFolder);
+    Path oldIndexPath = Path.of(indexFolder);
     indexFolder += "." + info.md5;
-    this.indexFolderPath = Paths.get(indexFolder);
-    Files.move(oldIndexPath, this.indexFolderPath);
+    this.indexFolderPath = Path.of(indexFolder);
+    if (!checkFileExist(this.indexFolderPath)) {
+      Files.move(oldIndexPath, this.indexFolderPath);
+    }
     return indexFolder;
   }
